+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - Extend `warp.fem.interpolate` API to be more flexible, decoupling interpolation location from storage definition
   ([GH-1091](https://github.com/NVIDIA/warp/issues/1091)).
 - Implement kernel caching for `wp.map()` to speed up repeated calls with the same function and similar input arguments ([GH-1108](https://github.com/NVIDIA/warp/issues/1108)).
+- Support converting between vector, matrix, and scalar types in `array.view()` ([GH-1112](https://github.com/NVIDIA/warp/issues/1112)).
 
 ### Fixed